pyproject.toml
[project]
name="dim"
version = "0.1.0"
description="Data Integrity Monitoring"
authors = [
    {name = "Alekhya Kommasani", email="akommasani@mozilla.com"},
    {name = "Krzysztof Ignasiak", email="kignasiak@mozilla.com"},
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    'attrs==22.1.0',
    'cattrs==22.2.0',
    "click==8.1.3",
    'db-dtypes>=1.0.4',
    'google-cloud-bigquery>=3.5.0',
    "Jinja2==3.1.2",
    'slack_sdk>=3.19.2',
    'tabulate==0.9.0',
    'pandas>=1.5.1',
    "pyyaml==6.0",
]

[project.optional-dependencies]
testing = [
  "black==22.10.0",
  "flake8==5.0.4",
  "isort==5.10.1",
  "mypy==0.982",
  "pytest>=7.1.3",
  "pytest-cov==4.0.0",
]

[project.urls]
repository = "https://github.com/mozilla/dim"
Homepage = "https://github.com/mozilla/dim"

[project.scripts]
dim = "dim.cli:cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["dim"]

[tool.black]
line-length = 100
target-version = ['py310',]
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
line_length = 100
